Refactoring e adicao de validacoes no MarcaController

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,15 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+    /** 
+     * Cria um atributo baseado no objeto injetado
+     * 
+     * @param \App\Models\Marca $marca
+    */
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +23,8 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marcas = Marca::all();
+        // $marcas = Marca::all();
+        $marcas = $this->marca->all();
         return $marcas;
     }
 
@@ -36,18 +46,23 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
+        // $marca = Marca::create($request->all());
+        $marca = $this->marca->create($request->all());
         return $marca;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show(int $id)
     {
+        $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -66,17 +81,19 @@ class MarcaController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, int $id)
     {
-        /*
-        print_r($request->all());// dados atualizados
-        echo '<hr>';
-        print_r($marca->getAttributes());// dados antigos
-        */
+        // $marca->update($request->all());
+        $marca = $this->marca->find($id);
 
+        if ($marca === null) {
+            return [
+                'erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'
+            ];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -84,11 +101,18 @@ class MarcaController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy(int $id)
     {
+        $marca = $this->marca->find($id);
+
+        if ($marca === null) {
+            return [
+                'erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe.'
+            ];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
